Shuffle modules around for better breakdown
